rs: add workaround for mine not colliding with Conrad in level 5.2
This happens in room 50 (and 51). piege[79] is the mine and piege[154]
is the launcher. It seems the obj/pge are missing the triggers as the
same glitch happens if playing the game under dosbox. Room 7 where
there is the same mine launcher does not have the problem and collides
well with Conrad when falling.
Interestingly, the Amiga version does not have this mine in that room.